reorganized metric utils slightly
