+++ credoai/utils/metric_constants.py
@@ -1,4 +1,3 @@
-from credoai.utils.metric_utils import STANDARD_CONVERSIONS, humanize_metric_name
 from credoai.utils.metrics import (
     equal_opportunity_difference, false_discovery_rate, false_omission_rate
 )
@@ -40,8 +39,31 @@ FAIRNESS_METRICS['equal_opportunity_difference'] = equal_opportunity_difference
 # Define functions that require probabilities ***
 PROBABILITY_METRICS = {"average_precision_score", "roc_auc_score"}
 
-# collate all metrics
+# *** Define Alternative Naming ***
+METRIC_EQUIVALENTS = {
+    'false_positive_rate': ['fpr', 'fallout_rate'],
+    'false_negative_rate': ['fnr', 'miss_rate'], 
+    'false_discovery_rate': ['fdr'], 
+    'true_positive_rate': ['tpr', 'recall_score', 'sensitivity', 'hit_rate'], 
+    'true_negative_rate': ['tnr', 'specificity'], 
+    'demographic_parity_difference': ['statistical_parity', 'demographic_parity'], 
+    'demographic_parity_ratio': ['disparate_impact'], 
+    'average_odds_difference': ['average_odds'], 
+    'equal_opportunity_difference': ['equal_opportunity'], 
+    'equalized_odds_difference': ['equalized_odds']
+}
+
+STANDARD_CONVERSIONS = {}
+for standard, equivalents in METRIC_EQUIVALENTS.items():
+    conversions = {name: standard for name in equivalents}
+    STANDARD_CONVERSIONS.update(conversions)
+    
+# *** collate all metrics ***
 ALL_METRICS = list(BINARY_CLASSIFICATION_METRICS.keys()) + \
     list(FAIRNESS_METRICS.keys()) + list(STANDARD_CONVERSIONS.keys())
 
+# humanize metric names
+def humanize_metric_name(metric):
+    return ' '.join(metric.split('_')).title()
 ALL_METRICS_HUMANIZED = [humanize_metric_name(m) for m in ALL_METRICS]
+
