+++ credoai/utils/metric_utils.py
@@ -1,25 +1,9 @@
 import re
 from credoai.utils.common import remove_suffix
-
-# *** Define Alternative Naming ***
-
-METRIC_EQUIVALENTS = {
-    'false_positive_rate': ['fpr', 'fallout_rate'],
-    'false_negative_rate': ['fnr', 'miss_rate'], 
-    'false_discovery_rate': ['fdr'], 
-    'true_positive_rate': ['tpr', 'recall_score', 'sensitivity', 'hit_rate'], 
-    'true_negative_rate': ['tnr', 'specificity'], 
-    'demographic_parity_difference': ['statistical_parity', 'demographic_parity'], 
-    'demographic_parity_ratio': ['disparate_impact'], 
-    'average_odds_difference': ['average_odds'], 
-    'equal_opportunity_difference': ['equal_opportunity'], 
-    'equalized_odds_difference': ['equalized_odds']
-}
-
-STANDARD_CONVERSIONS = {}
-for standard, equivalents in METRIC_EQUIVALENTS.items():
-    conversions = {name: standard for name in equivalents}
-    STANDARD_CONVERSIONS.update(conversions)
+from credoai.utils.metric_constants import (
+    BINARY_CLASSIFICATION_METRICS, FAIRNESS_METRICS, 
+    METRIC_EQUIVALENTS, STANDARD_CONVERSIONS
+)
 
 def standardize_metric_name(metric):
     # standardize
@@ -29,5 +13,14 @@ def standardize_metric_name(metric):
     standard = remove_suffix(remove_suffix(standard, '_difference'), '_parity')
     return STANDARD_CONVERSIONS.get(standard, standard)
 
-def humanize_metric_name(metric):
-    return ' '.join(metric.split('_')).title()
+def list_metrics():
+    metrics = {'performance': list(BINARY_CLASSIFICATION_METRICS.keys()) ,
+               'fairness': list(FAIRNESS_METRICS.keys())}
+    for key in metrics.keys():
+        names = metrics[key]
+        for val in names:
+            equivalents = METRIC_EQUIVALENTS.get(val)
+            if equivalents is not None:
+                metrics[key] += equivalents
+    metrics = {k: sorted(v) for k, v in metrics.items()}
+    return metrics
